feat: implement advanced GSAP scroll-triggered animations and state management for the services process section
