+++ src/components/MainContent/Landing/shared/index.ts
@@ -2,5 +2,6 @@ export { default as Navbar } from './Navbar/Navbar';
 export { default as CTAButton } from './CTAButton/CTAButton';
 export * from './animations/entranceAnimations';
 export * from './animations/servicesEntrance';
+export * from './animations/useServicesScrollDetection';
 export * from './fonts/fonts';
 export { default as CustomCursor } from './CustomCursor/CustomCursor';
